+++ scopes-macros/src/hierarchy/mod.rs
@@ -1,0 +1,4 @@
+mod hierarchy;
+pub(self) mod inclusion_graph;
+
+pub use hierarchy::*;